my컴포넌트분리
