+++ lib/ui/pages/other/my_page/wigets/my_tab_bar.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_blog/ui/pages/other/my_page/my_interest_author_view_model.dart';
 import 'package:flutter_blog/ui/pages/other/my_page/my_interest_webtoon_view_model.dart';
-import 'package:flutter_blog/ui/pages/other/my_page/wigets/my_comment.dart';
-import 'package:flutter_blog/ui/pages/other/my_page/wigets/my_interest_author.dart';
-import 'package:flutter_blog/ui/pages/other/my_page/wigets/my_interest_webtoon.dart';
-import 'package:flutter_blog/ui/pages/other/my_page/wigets/my_recently_view.dart';
+import 'package:flutter_blog/ui/pages/other/my_page/wigets/my_comment_page/my_comment.dart';
+import 'package:flutter_blog/ui/pages/other/my_page/wigets/my_interest_author_page/my_interest_author.dart';
+import 'package:flutter_blog/ui/pages/other/my_page/wigets/my_interest_webtoon_page/my_interest_webtoon.dart';
+import 'package:flutter_blog/ui/pages/other/my_page/wigets/my_recent_page/my_recently_view.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 class MyTabBar extends ConsumerWidget {
